requirements.txt
numpy==2.1.1
opencv_python==4.10.0.84
openvino==2024.3.0
openvino_dev==2024.3.0
Pillow==10.4.0
PyYAML==6.0.2
PyYAML==6.0.2
ultralytics==8.2.93
